+++ simclr/modules/__init__.py
@@ -1,7 +1,6 @@
 from .nt_xent import NT_Xent
-from .onnpu_plus_ntxent import  PU_plus_NTXent
 from .logistic_regression import LogisticRegression
 from .lars import LARS
 from .resnet import get_resnet
 from .gather import GatherLayer
-from .triplet_loss import MedianTripletHead, SmoothTripletHead, TripletNNPULoss
+from .new_losses import *
